Revert "hwc patch"
