+++ shims/libexynoscamera/Android.mk
@@ -14,12 +14,7 @@
 
 LOCAL_PATH := $(call my-dir)
 
-
 include $(CLEAR_VARS)
-
-LOCAL_SRC_FILES := \
-    CameraParameters.cpp \
-    Fence.cpp
 
 LOCAL_SRC_FILES := libexynoscamera.cpp
