--- shims/libexynoscamera/CameraParameters.h
@@ -1,94 +1,0 @@
-namespace android
-{
-
-class CameraParameters
-{
-public:
-    static const char PIXEL_FORMAT_YUV420SP_NV21[];
-    static const char EFFECT_CARTOONIZE[];
-    static const char EFFECT_POINT_RED_YELLOW[];
-    static const char EFFECT_POINT_GREEN[];
-    static const char EFFECT_POINT_BLUE[];
-    static const char EFFECT_VINTAGE_COLD[];
-    static const char EFFECT_VINTAGE_WARM[];
-    static const char EFFECT_WASHED[];
-    static const char ISO_AUTO[];
-    static const char ISO_NIGHT[];
-    static const char ISO_SPORTS[];
-    static const char ISO_6400[];
-    static const char ISO_3200[];
-    static const char ISO_1600[];
-    static const char ISO_800[];
-    static const char ISO_400[];
-    static const char ISO_200[];
-    static const char ISO_100[];
-    static const char ISO_80[];
-    static const char ISO_50[];
-    static const char KEY_SUPPORTED_METERING_MODE[];
-    static const char METERING_CENTER[];
-    static const char METERING_MATRIX[];
-    static const char METERING_SPOT[];
-    static const char METERING_OFF[];
-    static const char KEY_DYNAMIC_RANGE_CONTROL[];
-    static const char KEY_SUPPORTED_PHASE_AF[];
-    static const char KEY_PHASE_AF[];
-    static const char KEY_SUPPORTED_RT_HDR[];
-    static const char KEY_RT_HDR[];
-};
-	
-};
-/*
- * Copyright (C) 2017-2018 TeamNexus
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace android
-{
-
-class CameraParameters
-{
-public:
-	static const char PIXEL_FORMAT_YUV420SP_NV21[];
-	static const char EFFECT_CARTOONIZE[];
-	static const char EFFECT_POINT_RED_YELLOW[];
-	static const char EFFECT_POINT_GREEN[];
-	static const char EFFECT_POINT_BLUE[];
-	static const char EFFECT_VINTAGE_COLD[];
-	static const char EFFECT_VINTAGE_WARM[];
-	static const char EFFECT_WASHED[];
-	static const char ISO_AUTO[];
-	static const char ISO_NIGHT[];
-	static const char ISO_SPORTS[];
-	static const char ISO_6400[];
-	static const char ISO_3200[];
-	static const char ISO_1600[];
-	static const char ISO_800[];
-	static const char ISO_400[];
-	static const char ISO_200[];
-	static const char ISO_100[];
-	static const char ISO_80[];
-	static const char ISO_50[];
-	static const char KEY_SUPPORTED_METERING_MODE[];
-	static const char METERING_CENTER[];
-	static const char METERING_MATRIX[];
-	static const char METERING_SPOT[];
-	static const char METERING_OFF[];
-	static const char KEY_DYNAMIC_RANGE_CONTROL[];
-	static const char KEY_SUPPORTED_PHASE_AF[];
-	static const char KEY_PHASE_AF[];
-	static const char KEY_SUPPORTED_RT_HDR[];
-	static const char KEY_RT_HDR[];
-};
-	
-};
