--- shims/libexynoscamera/Fence.cpp
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2018 TeamNexus
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-extern "C"
-{
-
-//
-// External deconstructor symbol got removed in P
-// Replaced with empty default implementation:
-//
-//     ~Fence() = default;
-//
-void *_ZN7android5FenceD1Ev(void *_this) {
-	return _this;
-}
-
-}
