+++ shims/libexynoscamera/libexynoscamera.h
@@ -1,0 +1,39 @@
+namespace android
+{
+
+class CameraParameters
+{
+public:
+    static const char PIXEL_FORMAT_YUV420SP_NV21[];
+    static const char EFFECT_CARTOONIZE[];
+    static const char EFFECT_POINT_RED_YELLOW[];
+    static const char EFFECT_POINT_GREEN[];
+    static const char EFFECT_POINT_BLUE[];
+    static const char EFFECT_VINTAGE_COLD[];
+    static const char EFFECT_VINTAGE_WARM[];
+    static const char EFFECT_WASHED[];
+    static const char ISO_AUTO[];
+    static const char ISO_NIGHT[];
+    static const char ISO_SPORTS[];
+    static const char ISO_6400[];
+    static const char ISO_3200[];
+    static const char ISO_1600[];
+    static const char ISO_800[];
+    static const char ISO_400[];
+    static const char ISO_200[];
+    static const char ISO_100[];
+    static const char ISO_80[];
+    static const char ISO_50[];
+    static const char KEY_SUPPORTED_METERING_MODE[];
+    static const char METERING_CENTER[];
+    static const char METERING_MATRIX[];
+    static const char METERING_SPOT[];
+    static const char METERING_OFF[];
+    static const char KEY_DYNAMIC_RANGE_CONTROL[];
+    static const char KEY_SUPPORTED_PHASE_AF[];
+    static const char KEY_PHASE_AF[];
+    static const char KEY_SUPPORTED_RT_HDR[];
+    static const char KEY_RT_HDR[];
+};
+	
+};
